+++ src/main/java/univ/kgu/carely/domain/meet/service/impl/MemoServiceImpl.java
@@ -2,7 +2,7 @@ package univ.kgu.carely.domain.meet.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import univ.kgu.carely.domain.meet.repository.MemoRepository;
+import univ.kgu.carely.domain.meet.repository.memo.MemoRepository;
 import univ.kgu.carely.domain.meet.service.MemoService;
 import univ.kgu.carely.domain.member.service.MemberService;
